Use the new scenario class
