+++ src/hisp/__init__.py
@@ -1,3 +1,4 @@
-from .helpers import PulsedSource, Scenario
+from .helpers import PulsedSource
 
 from .h_transport_class import CustomProblem
+from .scenario import Scenario, Pulse
